refactor: 예산 추천 가이드 데이터 가공 로직 분리 및 적용 #99
- transformBudgetGuideData 함수를 analysis-transform.ts로 분리
